+++ src/config.rs
@@ -830,6 +830,7 @@ impl Config {
 }
 
 #[cfg(test)]
+#[allow(clippy::result_large_err)]
 mod tests {
     use super::*;
     use figment::Jail;